Move it on_interation ralated code to proper module
