+++ src/python/mod.rs
@@ -14,6 +14,9 @@ mod impl_default_py;
 pub(crate) mod io;
 mod module_py;
 
+/// Python interface for the `on_iteration` callback.
+pub mod on_iteration;
+
 // compile this module if no local blas/lapack library
 // has been specified, and we want to use the python/scipy
 // version instead.  sdp_pyblas is defined in build.rs
